src/requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
antlr4-python3-runtime==4.8
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asteroid-filterbanks==0.4.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
autopage==0.5.0
av==9.1.1
backcall==0.2.0
backports.cached-property==1.0.1
beautifulsoup4==4.10.0
bleach==5.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.0
colorama==0.4.4
colorlog==6.6.0
cycler==0.11.0
debugpy==1.6.0
decorator==4.4.2
deepface==0.0.74
defusedxml==0.7.1
docopt==0.6.2
einops==0.3.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
facenet-pytorch==2.5.2
fastjsonschema==2.15.3
ffmpeg-python==0.2.0
filelock==3.6.0
Flask==2.1.1
flatbuffers==1.12
fonttools==4.32.0
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
fvcore==0.1.5.post20220305
gast==0.4.0
gdown==4.4.0
google-auth==2.6.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.34.1
h11==0.12.0
h2==4.1.0
h5py==3.1.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
huggingface-hub==0.5.1
hyperframe==6.0.1
HyperPyYAML==1.0.0
idna==3.3
ImageHash==4.2.1
imageio==2.16.2
imageio-ffmpeg==0.4.5
imbalanced-learn==0.9.0
importlib-metadata==4.11.3
iopath==0.1.9
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
julius==0.2.6
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.8.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
librosa==0.8.0
llvmlite==0.36.0
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
moviepy==1.0.3
mpmath==1.2.1
mtcnn==0.1.1
multidict==6.0.2
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8
nltk==3.7
norbert==0.2.1
notebook==6.4.10
numba==0.53.1
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
optuna==2.10.0
packaging==21.3
pandas==1.4.2
pandasql==0.7.3
pandocfilters==1.5.0
parameterized==0.8.1
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pooch==1.6.0
portalocker==2.4.0
prettytable==3.2.0
primePy==1.3
proglog==0.1.9
prometheus-client==0.14.0
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio @ https://github.com/pyannote/pyannote-audio/archive/develop.zip
pyannote.core==4.4
pyannote.database==4.1.3
pyannote.metrics==3.2
pyannote.pipeline==2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.11.2
pyparsing==3.0.7
pyperclip==1.8.2
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.6.0
pytorch-metric-learning==1.3.0
pytorchvideo==0.1.5
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
retina-face==0.0.10
rfc3986==1.5.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.49
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
shellingham==1.4.0
simplejson==3.17.6
singledispatchmethod==1.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3.2
spectralcluster==0.2.5
speechbrain==0.5.11
spleeter==2.3.0
SQLAlchemy==1.4.35
stack-data==0.2.0
stevedore==3.5.0
sympy==1.10.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
textblob==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.4.8
tinydb==4.7.0
tinydb-serialization==2.1.0
tokenizers==0.11.6
torch==1.11.0
torch-audiomentations @ git+https://github.com/asteroid-team/torch-audiomentations@7482d0451c93c7bb88b5c6330a8dfd0c242fc600
torch-pitch-shift==1.2.2
torchaudio==0.11.0
torchmetrics==0.7.3
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.18.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
wrapt==1.12.1
xmltodict==0.12.0
yacs==0.1.8
yarl==1.7.2
zipp==3.8.0
